requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pymongo[srv]==4.6.0
playwright==1.40.0
motor==3.3.2
pydantic==2.5.0
loguru==0.7.2
psutil>=5.9.0
certifi>=2023.5.7
urllib3>=1.26.0
pyOpenSSL>=23.0.0